+++ HISTORY.md
@@ -1,5 +1,6 @@
-## v2.3.25
+## v2.3.100
 * First batch of expired commercial licenses have been banned for use.
+* Issue #139 - New `.RunUnsafeAsync` run helper that runs a query on the server and returns the raw `Response` object in an unsafe way with minimal client-side processing. The unbridled use of this method can lead to memory leaks on both the client and server. Use with extreme caution. This method should only be used in well-understood and advanced scenarios.
 * **POSSIBLE BREAKING CHANGE:** `.RunResult` for `Result` (DML) type responses that correspond to inserts, deletes, updates has been renamed to `.RunWrite`. Please use the new method name instead. **Note:** The `.RunResult<T>` for `T` type responses will not be obsoleted. The reason for the former `.RunResult` method rename is to avoid confusion between `.RunResult` and `.RunResult<T>`. The obsoleted methods produce compiler warnings in this version. If no compiler warnings are produced, no changes are necessary. The obsolete methods will be removed in the next major RethinkDB driver release (2.4).
 
 ## v2.3.24